samsun_project/parser/requirements.txt
yt-dlp
openai-whisper
torch
torchaudio
transformers